+++ src/MixtureModel.cpp
@@ -28,7 +28,7 @@ MixtureModel::~MixtureModel() {
 
 float MixtureModel::get_log_prob(const Sample* s) {
     float l_prob = -std::numeric_limits<double>::max();
-    for (unsigned int c=0; c<m_models.size(); c++) {
+    for (size_t c=0; c<m_models.size(); c++) {
         log_sum_log(l_prob, log(m_mixing_probs[c]) + m_models[c]->get_log_prob(s));
         //prob += m_mixing_probs[c] * exp(m_models[c]->get_log_prob(s));
     }
@@ -71,12 +71,12 @@ void MixtureModel::e_step(const ModelData& data) {
     //	cerr <<"E-STEP" << endl;
     for (int i=0; i<data.get_number_of_samples(); i++) {
         float total_sample_posteriors=0;
-        for (unsigned int c=0; c<m_models.size(); c++) {
+        for (size_t c=0; c<m_models.size(); c++) {
             m_mixture_posteriors[c][i] = m_mixing_probs[c] * exp(m_models[c]->get_log_prob(data.get_sample(i)));
             total_sample_posteriors += m_mixture_posteriors[c][i];
         }
         //		cerr << "\tmix_posteriors [" << i << "] :: ";
-        for (unsigned int c=0; c<m_models.size(); c++) {
+        for (size_t c=0; c<m_models.size(); c++) {
             m_mixture_posteriors[c][i] /= total_sample_posteriors;
             //			cerr << "\t" << m_mixture_posteriors[c][i];
         }
@@ -86,7 +86,7 @@ void MixtureModel::e_step(const ModelData& data) {
 
 float MixtureModel::m_step(const ModelData& data) {
     //	cerr <<"M-STEP" << endl;
-    for (unsigned int c=0; c<m_models.size(); c++) {
+    for (size_t c=0; c<m_models.size(); c++) {
         m_models[c]->learn(data, m_mixture_posteriors[c]);
         //		m_models[c]->print(cerr);
         m_mixing_probs[c] = 0;
@@ -100,7 +100,7 @@ float MixtureModel::m_step(const ModelData& data) {
 
 void MixtureModel::print(ostream& out) {
     out << endl;
-    for (unsigned int i=0; i<m_models.size(); i++) {
+    for (size_t i=0; i<m_models.size(); i++) {
         out << i << " :: mixing prob:" << m_mixing_probs[i] << "\t";
         m_models[i]->print(out);
     }
@@ -108,7 +108,7 @@ void MixtureModel::print(ostream& out) {
 }
 
 void MixtureModel::print_summary(ostream& out) {
-    for (unsigned int i=0; i<m_models.size(); i++) {
+    for (size_t i=0; i<m_models.size(); i++) {
         out << m_mixing_probs[i] << "\t";
         m_models[i]->print_summary(out);
         out << "\t";
